Version 1.2
The program was made to view 3D models in Godot 4.3, analyze any project and directory outside of this project, and display a list of models that are in the project or directory on the computer.
Implemented
Viewing files in the “glb” and “gltf” formats.
Ability to partially view materials
Ability to view the model from all angles and distances
Fixed.
Bug with zooming in and out while scrolling in the list
Added
Auto-rotation on Space
Saving an open project and model and view

Translated with DeepL.com (free version)